Split UnusedPrivateMemberTest in multiple files
Also preped all test methods with UnusedPrivateMember_ to be easy to filter the 25 tests inside the IDE
